feat: Support namespace in monitor
