+++ cmd/ui/v1alpha3/main.go
@@ -29,7 +29,7 @@ func main() {
 	http.Handle("/katib/", http.StripPrefix("/katib/", frontend))
 
 	http.HandleFunc("/katib/fetch_hp_jobs/", kuh.FetchAllHPJobs)
-	http.HandleFunc("/katib/fetch_nas_jobs/", kuh.FetchNASJobs)
+	http.HandleFunc("/katib/fetch_nas_jobs/", kuh.FetchAllNASJobs)
 	http.HandleFunc("/katib/submit_yaml/", kuh.SubmitYamlJob)
 	http.HandleFunc("/katib/submit_hp_job/", kuh.SubmitParamsJob)
 	http.HandleFunc("/katib/submit_nas_job/", kuh.SubmitParamsJob)
