+++ pkg/ui/v1alpha3/frontend/src/actions/nasMonitorActions.js
@@ -24,7 +24,8 @@ export const FETCH_NAS_JOB_INFO_REQUEST = "FETCH_NAS_JOB_INFO_REQUEST";
 export const FETCH_NAS_JOB_INFO_SUCCESS = "FETCH_NAS_JOB_INFO_SUCCESS";
 export const FETCH_NAS_JOB_INFO_FAILURE = "FETCH_NAS_JOB_INFO_FAILURE";
 
-export const fetchNASJobInfo = (experimentName) => ({
+export const fetchNASJobInfo = (experimentName, namespace) => ({
     type: FETCH_NAS_JOB_INFO_REQUEST,
-    experimentName
+    experimentName,
+    namespace
 })
